refactor: move tabs to their own component
